Fix missing ZSTD_CLEVEL_DEFAULT definition
When libbacktrace is compiled with ZSTD pre v1.3.6, the define
ZSTD_CLEVEL_DEFAULT was expected to be set by the application.

After 1.3.6, if the application does not set this define, the zstd.h
header sets it to 3.

This commit sets ZSTD_CLEVEL_DEFAULT to 3 before including zstd.h, which
allows gdb to build on older ZSTD libraries, as well as keep a
consistent compression level when build with newer libraries.